Code style fixes, some renaming
